Failed attempt to draw general triangles
I'm gonna try implementing the general triangles following 
http://www.hugi.scene.org/online/coding/hugi%2017%20-%20cotriang.htm
more closely. The other one 
http://www.sunshine2k.de/coding/java/TriangleRasterization/TriangleRasterization.html#sunbresenhamarticle
is cool, but calculates the inverse slope of the longer edge twice per triangle (One per flat bottom and other for flat top).
Also, having to take care of the order of the vertex for each particular case is giving me problems and complicating the code unnecessarily. I don't like that. I could solve that implementing a more general drawing functions (In which it's not necessary to draw from left to right), tough.